pyproject.toml
[project]
name = "lodfunnel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.2",
    "click>=8.3.0",
    "ipykernel>=7.0.1",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "nltk>=3.9.3",
    "ollama>=0.6.1",
    "pandas>=2.3.3",
    "qlever>=0.5.33",
    "rdflib>=7.2.1",
    "ruamel-yaml>=0.18.15",
    "sentence-transformers>=5.3.0",
    "sparqlwrapper>=2.0.0",
    "tabulate>=0.10.0",
    "torch>=2.10.0",
    "torchmetrics>=1.8.2",
    "tqdm>=4.67.1",
    "transformers>=5.2.0",
]
